MAINT Relocate functions in pywasmcross (#3464)
