+++ pyodide-build/pyodide_build/tests/test_pywasmcross.py
@@ -8,7 +8,6 @@ from pyodide_build.pywasmcross import handle_command_generate_args  # noqa: E402
 from pyodide_build.pywasmcross import replay_f2c  # noqa: E402
 from pyodide_build.pywasmcross import (
     calculate_exports,
-    environment_substitute_args,
     get_cmake_compiler_flags,
     get_library_output,
 )
@@ -180,25 +179,6 @@ def test_conda_unsupported_args():
 
     assert generate_args("gcc -c test.o -Wl,--sysroot=/ -o test.so", args) == (
         "emcc -c test.o -o test.so"
-    )
-
-
-def test_environment_var_substitution(monkeypatch):
-    monkeypatch.setenv("PYODIDE_BASE", "pyodide_build_dir")
-    monkeypatch.setenv("BOB", "Robert Mc Roberts")
-    monkeypatch.setenv("FRED", "Frederick F. Freddertson Esq.")
-    monkeypatch.setenv("JIM", "James Ignatius Morrison:Jimmy")
-    args = environment_substitute_args(
-        {
-            "ldflags": '"-l$(PYODIDE_BASE)"',
-            "cxxflags": "$(BOB)",
-            "cflags": "$(FRED)",
-        }
-    )
-    assert (
-        args["cflags"] == "Frederick F. Freddertson Esq."
-        and args["cxxflags"] == "Robert Mc Roberts"
-        and args["ldflags"] == '"-lpyodide_build_dir"'
     )
